smart: Refactor and split into libbd_smartmontools
Refactoring in preparation for a second implementation based
on libatasmart.
