+++ Makefile.am
@@ -44,8 +44,8 @@ if !WITH_NVME
 DISTCHECK_CONFIGURE_FLAGS += --without-nvme
 endif
 
-if !WITH_SMART
-DISTCHECK_CONFIGURE_FLAGS += --without-smart
+if !WITH_SMARTMONTOOLS
+DISTCHECK_CONFIGURE_FLAGS += --without-smartmontools
 endif
 
 if !WITH_SWAP
@@ -72,7 +72,7 @@ MAINTAINERCLEANFILES = Makefile.in aclocal.m4 config.guess config.sub \
     configure depcomp install-sh ltmain.sh missing py-compile compile ar-lib \
     m4/*.m4
 
-LIBDIRS = src/utils/.libs:src/plugins/.libs:src/plugins/fs/.libs:src/plugins/nvme/.libs:src/lib/.libs
+LIBDIRS = src/utils/.libs:src/plugins/.libs:src/plugins/fs/.libs:src/plugins/nvme/.libs:src/plugins/smart/.libs:src/lib/.libs
 GIDIR = src/lib
 
 if WITH_PYTHON3
