+++ data/conf.d/00-default.cfg
@@ -46,7 +46,7 @@ sonames=libbd_nvme.so.3
 sonames=libbd_part.so.3
 
 [smart]
-sonames=libbd_smart.so.3
+sonames=libbd_smartmontools.so.3
 
 [swap]
 sonames=libbd_swap.so.3
