+++ src/plugins/smart/smart.h
